requirements.txt
amazon-textract-caller==0.2.3
amazon-textract-response-parser==1.0.2
asgiref==3.8.1
asttokens==2.4.1
boto3==1.34.105
botocore==1.34.105
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.6
django-crispy-forms==2.1
django-extensions==3.2.3
exceptiongroup==1.2.1
executing==2.0.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
idna==3.7
ipython==8.24.0
jedi==0.19.1
jmespath==1.0.1
marshmallow==3.21.2
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13